requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
discord.py==2.5.2
frozenlist==1.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
httplib2==0.22.0
idna==3.10
multidict==6.6.4
oauth2client==4.1.3
oauthlib==3.3.1
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
